// src/pages/main-menu/order/user/menu/subscriber/components/BSSteps/model/interface.ts
// Generated by https://quicktype.io

export interface SubsPlanByCatgList {
  type: "C" | "S" | "O";
  id: number | null;
  name: string;
  nodeId: string;
  parentCatgId: string;
  parentCatgIdSort: number;
  offerType: null | string;
  salePrice: null | string;
  rentListPrice: null;
  cycleQuantity: null;
  timeUnit: null;
  expDate: null;
  indepProdSpecId: number | null;
  servType: null | string;
  offerVerId: number | null;
  comments: null | string;
  paidFlag: "N" | "Y" | null;
  seq: number;
  servTypeReal: null | string;
  indepProdSpecIdReal: number | null;
  offerName: null | string;
  indepProdSpecName: null | string;
  priority: number;
  offerCatgMemId: number;
  offerCatgId: number;
  offerId: number | null;
  childOfferCatgId: number | null;
  spId: number;
  parentSeq: null;
  attrId: null;
  defaultValue: null;
}

// Generated by https://quicktype.io

export interface SubsPlanIDList {
  subsPlanId: number;
  flag: boolean;
  spId: number;
  indepProdSpecId: number;
}